update to liinking users to specific metrics
