+++ renderpup/server/controllers/metrics.js
@@ -47,7 +47,7 @@ metricsController.timeToFirstByte = async (req, res, next) => {
 
 metricsController.getDatabaseData = async (req, res, next) => {
   //Selects all data from the metrics table and attaches the rows to the res.locals object
-  const data = await db.query(`SELECT * FROM metrics WHERE url='${req.body.url}/'`)
+  const data = await db.query(`SELECT * FROM metrics WHERE url='${req.body.url}/' AND user_id=${req.cookies.userId}`)
   res.locals.databaseData = data.rows
   
   next()
@@ -55,7 +55,7 @@ metricsController.getDatabaseData = async (req, res, next) => {
 
 metricsController.getUrls = async (req, res, next) => {
   //Selects all urls from the metrics table and attaches the rows to the res.locals object
-  const data = await db.query('SELECT url FROM metrics')
+  const data = await db.query(`SELECT url FROM metrics WHERE user_id = '${req.cookies.userId}'`)
   res.locals.urls = data.rows
   
   next()
@@ -68,9 +68,9 @@ metricsController.saveMetrics = async (req, res, next) => {
   // const diagnostics = res.locals.diagnostics
   const opportunities = res.locals.opportunities
   res.locals.metrics.date = new Date()
-  await db.query('INSERT INTO metrics (url, ttfb, fcp, lcp, nsl, bs) ' + `VALUES ('${url}', ${ttfb}, ${fcp}, ${lcp}, ${nsl}, '${JSON.stringify(bs)}')`);
+  await db.query('INSERT INTO metrics (url, user_id, ttfb, fcp, lcp, nsl, bs) ' + `VALUES ('${url}', ${req.cookies.userId}, ${ttfb}, ${fcp}, ${lcp}, ${nsl}, '${JSON.stringify(bs)}')`);
   
-  // await db.query('INSERT INTO diagnostics (performance_score, diagnostics_info) ' + `VALUES (${performanceScore}, '${opportunities}')`)
+  // await db.query('INSERT INTO diagnostics (performance_score, diagnostics_info, url) ' + `VALUES (${performanceScore}, '${opportunities}', '${url}')`)
   next()
 };
