+++ renderpup/server/routers/api.js
@@ -19,11 +19,11 @@ router.post('/urls', metricsController.getDatabaseData, (req, res) => {
   return res.status(200).json({data: res.locals.databaseData});
 })
 
-router.post('/login', userController.verifyUser, (req, res) => {
+router.post('/login', userController.verifyUser, userController.createUserIdCookie, (req, res) => {
   return res.status(200).json(res.locals.passwordMatches);
 })
 
-router.post('/signup', userController.createUser, (req, res) => {
+router.post('/signup', userController.createUser, userController.createUserIdCookie, (req, res) => {
   return res.status(200).json(res.locals.userCreated);
 })
